requirements.txt
# Dependencias principales
osmnx>=1.9.0
networkx>=3.1
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
geopandas>=0.14.0
folium>=0.14.0
scikit-learn>=1.3.0

# Análisis y procesamiento
scipy>=1.11.0
shapely>=2.0.0

shap>=0.43.0
xgboost>=2.0.0

# Visualizaciones avanzadas
plotly>=5.14.0
contextily>=1.3.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0

# Reportes
fpdf2>=2.7.0
reportlab>=4.0.0

# Utilidades
pyyaml>=6.0
tqdm>=4.65.0
